web/requirements.txt
asgiref==3.3.1
Babel==2.9.0
Click==7.1.2
croniter==0.3.36
Django==3.1.4
django-model-utils==4.0.0
django-redis==4.12.1
django-rq==2.4.0
rq-scheduler==0.10.0
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
rq==1.7.0
six==1.15.0
sqlparse==0.4.1
psycopg2==2.8.6
gunicorn==20.0.4
whitenoise==5.2.0
django-fullurl==1.1
requests==2.25.0
aiosmtplib==1.1.4
beautifulsoup4==4.9.3
pycparser==2.20
soupsieve==2.0.1
djangorestframework==3.12.2
pyjwt==1.7.1
django-cors-headers==3.5.0
django-filter==2.4.0